Using single file.
