+++ packages/gallery/buildScripts/buildRuntimeValidateFunctionFromTypes.js
@@ -1,31 +1,25 @@
+const TJS = require('typescript-json-schema');
 const path = require('path');
 const browserify = require('browserify');
 const fs = require('fs');
 const Ajv = require('ajv');
 
-const getSchemaFromTypes = require('./generateJSONSchemaFromTypes');
-
-function writeES5StandaloneValidateMethod() {
-  const galleryFolder = path.join(__dirname, '../src/components/gallery');
-  const sourceTypesFile = path.join(galleryFolder, 'galleryTypes.ts');
-  const code = buildValidationFunction(
-    getSchemaFromTypes(sourceTypesFile)
-  );
-  const tempFilePath = path.join(__dirname, 'temp.js');
-  fs.writeFileSync(tempFilePath, code);
-  const targetFilePath = path.join(
-    galleryFolder,
-    'typeValidator/standaloneValidateCode.js'
-  );
-  const fileWriter = fs.createWriteStream(targetFilePath);
-  browserify(tempFilePath, { standalone: 'nirnaor' })
+function writeES5StandaloneValidateMethod({
+  sourceTypesFile,
+  targetFile,
+  tempFile,
+}) {
+  const code = buildValidationFunction(getSchemaFromTypes(sourceTypesFile));
+  fs.writeFileSync(tempFile, code);
+  const fileWriter = fs.createWriteStream(targetFile);
+  browserify(tempFile, { standalone: 'nirnaor' })
     .transform('babelify', { global: true, presets: ['@babel/preset-env'] })
     .bundle()
     .pipe(fileWriter);
 
   fileWriter.on('finish', function () {
     console.log('finished writing the browserify file');
-    fs.rmSync(tempFilePath);
+    fs.rmSync(tempFile);
   });
 }
 
@@ -42,4 +36,36 @@ function buildValidationFunction(schema) {
   return moduleCode;
 }
 
-writeES5StandaloneValidateMethod();
+function getSchemaFromTypes(typesFileAbsolutePath) {
+  // optionally pass argument to schema generator
+  const settings = {
+    required: true,
+  };
+
+  // optionally pass ts compiler options
+  const compilerOptions = {
+    strictNullChecks: true,
+  };
+
+  // optionally pass a base path
+  // const basePath = "./my-dir";
+
+  const program = TJS.getProgramFromFiles(
+    [typesFileAbsolutePath],
+    compilerOptions
+    // basePath
+  );
+
+  // We can either get the schema for one file and one type...
+  const schema = TJS.generateSchema(program, 'StyleParams', settings);
+  return schema;
+}
+
+const galleryFolder = path.join(__dirname, '../src/components/gallery');
+const sourceTypesFile = path.join(galleryFolder, 'galleryTypes.ts');
+const tempFile = path.join(__dirname, 'temp.js');
+const targetFile = path.join(
+  galleryFolder,
+  'typeValidator/standaloneValidateCode.js'
+);
+writeES5StandaloneValidateMethod({ sourceTypesFile, targetFile, tempFile });
